Sync c-h nrpe.py file to fix nrpe files copy issue
Due to a change to PY3, the charmhelpers directory is not in the
CHARMDIR rather than in CHARMDIR/hooks. This broke the
copy_nrpe_checks() function in c-h. c-h has been updated, to look in
both CHARMDIR and CHARMDIR/hooks for the charmhelpers directory, and
this patchset includes the new function definition.

Change-Id: I5cd14bede21f53f196c630c7fdfcf521853e68da
Closes-Bug: #1796830